requirements.txt
# Google Gemini Gen AI SDK
google-generativeai>=0.3.0

# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# Async support
asyncio>=3.4.3

# Scientific computing (for validation and advanced math)
numpy>=1.24.0
scipy>=1.10.0
# HATA: sympy eksik - calculus modülünde kullanılıyor ama requirements'da yok
# sympy>=1.12  # COMMENT OUT edilmiş!

# Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging and utilities
python-json-logger>=2.0.7

